Introduce a parallel_backend_sycl_reduce_then_scan.h file to contain implementation
